requirements.txt
yt-dlp
requests
gradio
click
python-dotenv